Cutting parts out of other parts
Refactored so that we take proper notice of whether we've got a component or an occurrence (where before we kinda used them interchangeably).

Also refactored the parts module so there's now a common base class

Annoyingly, the cut functionality is bugged. If you remove the parent node it works fine, and if you manually edit the feature in the timeline (it'll be flagged as having an error) and then immediately confirm without changing anything, it'll be magically fixed